Do not link client before platform.
libplatform depends on libclient and thus linking with
platform also links with client.

However -- on linux, at least when using ld.gold (which is the default
linker on most platforms now), the order of linking static libraries
is important and cmake removes the seemingly duplicated libclient.a
that platform adds. Hence, by linking with "... client ... platform ...",
the resulting link command contains 'libclient.a ... libplatform.a',
in that order which causes an undefined reference in libplatform.a
for references to libclient.a.

By removing the mentions of 'client' when we link with platform,
the final order is swapped and things work link again.

Note that currently syngergy is not actually using libclient from
libplatform on linux (but only on OSX and windows) so that this problem
was not spotted before. However in subsequent patches I will add
this dependency for linux too and therefore need this patch.