Add protocol 1.6 client with msgs kMsgDMouseWarp (DMWA) and kMsgDLockScreen (DLCK)
Adds CClient::warpMouse, CClient::lockScreen, CClient::mouseWarp,
CServerProxy::mouseWarp, CServerProxy::warpMouse, CServerProxy::lockScreen,
and CBaseClientProxy::isLockedToScreen.
Removes CScreen::isLockedToScreen because that always returned false,
and removes CPrimaryClient::isLockedToScreen because that called
CScreen::isLockedToScreen.

The CServer::onMouseMoveSecondary was changed use relative mouse
coordinates when isLockedToScreen() returns true, rather than
when m_relativeMoves && isLockedToScreenServer() because m_relativeMoves
is reset to false anyway, and we're not interested in only what the
server thinks, but also what the clients think now: isLockedToScreen()
returns true when isLockedToScreenServer() is true, but also when
a client asked for a lock (see DLCK below).

The patch adds support for two new protocol messages:
DMWA : Mouse WArp
DLCK : screen LoCK

The latter only travels from client to server and requests
the server to be locked or unlocked (parameter 1 or 0 respectively).
While locked, the server sends this client only relative mouse
movements and does not detect edges.

DMWA takes two signed integers as parameter which represent the
relative offsets of the mouse pointer (the warp).
When sent from client to server, it is a request to warp the internal
coordinates that are used, after which the server sends it back
with the same values to indicate that it did just that.

When the client receives the latter, it calls CPlatformScreen::mouseWarp
which currently does nothing. Nor does the client currently send
these new messages; but they will be used in the next commit.