Screen lock bug fix.
This fixes a bug where the client would remain locked to its screen.

Note that if an application grabs the mouse then the synergy client does
not receive motion notify events anymore. This can result in not being
able anymore to leave the screen as soon as you press(ed) a mouse button
if that application is fullscreen. After this patch you can still leave
the screen then by either going through the title bar, or the task bar,
or pressing ScrLk twice - which now also unlocks the client (the first
press it locks the cursor to the client screen by authority of the
server, and the second press it unlocks that AND unlocks the lock
by authority of the client, overriding its authority).